+++ src/main/Java/com/accenture/flowershop/be/BusinessService/Flower/FlowerBusinessServiceImpl.java
@@ -27,8 +27,10 @@ public class FlowerBusinessServiceImpl implements FlowerBusinessService {
         }
     }
 
+
+
     @Override
-    public Flower updateFlowersNumber(Long id, int number) {
+    public Flower updateFlowersNumber(Long id, Long number) {
             Flower flower = flowerDAO.findFlowerById(id);
             flower.setNumber(number);
             return flower;
